+++ elements/src/components/pos-relations/pos-relations.tsx
@@ -24,7 +24,7 @@ export class PosRelations implements ResourceAware {
     const items = this.data.map(it => (
       <ion-item-group>
         <ion-item-divider>
-          <ion-label>{it.predicate}</ion-label>
+          <ion-label title={it.predicate}>{it.label}</ion-label>
         </ion-item-divider>
         {it.uris.map(uri => (
           <pos-rich-link uri={uri} />