+++ elements/CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - [pos-label](../docs/elements/components/pos-label): shorter fallback labels 
 - [pos-rich-link](../docs/elements/components/pos-rich-link): Redesigned to reduce clutter and duplicate information
+- [pos-relations](../docs/elements/components/pos-relations): shorter labels for predicates
+- [pos-reverse-relations](../docs/elements/components/pos-reverse-relations): shorter labels for predicates
 
 ## 0.20.0